Overview of project
Contains separate scripts that scrape CRAN website for packages or
use available.packages() function. (this part is a bit of a mess)
Also scrapes the archive for links.

Pulls down all .tar.gz files and processes them:

* calculate 3 different file hashes on tar.gz file
* unpack tar files
* create folder
* extract DESCRIPTION and copy to folder
* extract NAMESPACE and copy to folder
* copy info like hashes to INFO file in folder
* copy all man pages to folder
* delete tar.gz file

There is also a script to process the local files.

* extract filehashes and place them in sqlite table 'hashes'
* extract NAMESPACE and create entries for every import in sqlite table  'imports' and all exports in sqlite table 'exports'

There is also a file to create exports of hashes to 3 different csvs.